home_work12/requirements.txt
fastapi
sqlmodel
psycopg2
python-dotenv